+++ application/controllers/Player_Portfolio.php
@@ -19,8 +19,10 @@ class Player_Portfolio extends MY_Controller {
         $this->data['pagetitle'] = 'Player Portfolio';
         $this->data['pagebody'] = 'player_portfolio';
         $this->loadPlayerNamesToOptions();
-        $this->loadCurrentHoldings();
-        $this->loadTradingActivity();
+
+        $this->data['trading_activity'] = NULL;
+        $this->data['holdings'] = NULL;
+
         $this->render();
     }
 
@@ -37,37 +39,56 @@ class Player_Portfolio extends MY_Controller {
     }
 
     /**
-     * Load the current holdings table for the portfolio selected
+     * Routing function for the specific player
      */
-    function loadCurrentHoldings() {
-        $this->table->set_heading('Stock Code','Cash', 'Date');
-        $selectedPortfolio = $this->input->post('portfolio-select');
-        $this->data['playerName'] = $selectedPortfolio;
-
-        if (!empty($selectedPortfolio)) {
-            $portfolio = $this->PortfolioModel->getCurrentHoldings
-            ($selectedPortfolio);
-            foreach ($portfolio as $row) {
-                $this->table->add_row($row);
-            }
-        }
-        $this->data['holdings'] = $this->table->generate();
+    function getSpecificPortfolio() {
+        $player = $this->input->get('portfolio-select');
+        redirect("player-portfolio/$player");
     }
 
     /**
-     * Load the trading activity table for the portfolio selected.
+     * Displays a given portfolio with tables of trading activity and current holdings
+     * @param $player the portfolio player
      */
-    function loadTradingActivity() {
-        $this->table->set_heading('Stock Code', 'Transaction', 'Date');
-        $selectedPortfolio = $this->input->post('portfolio-select');
-        $this->data['playerName'] = $selectedPortfolio;
+    function displayPortfolio($player) {
+        $this->data['pagetitle'] = $player;
+        $this->data['pagebody'] = 'player_portfolio';
+        $this->loadPlayerNamesToOptions();
 
-        if (!empty($selectedPortfolio)) {
-            $portfolio = $this->PortfolioModel->getTradingActivity($selectedPortfolio);
-            foreach ($portfolio as $row) {
-                $this->table->add_row($row);
-            }
+        $tradingActivity = $this->PortfolioModel->getTradingActivity($player);
+        $this->data['trading_activity'] = $this->generateTradingActivityTable($tradingActivity);
+
+        $currentHoldings = $this->PortfolioModel->getCurrentHoldings($player);
+        $this->data['holdings'] = $this->generateCurrentHoldingsTable($currentHoldings);
+
+        $this->render();
+    }
+
+    /**
+     * Generates the trading activity table with a given player
+     * @param $player the portfolio player
+     * @return mixed The generated table
+     */
+    function generateTradingActivityTable($player) {
+        $this->table->set_heading('Stock Code', 'Transaction', 'Date');
+        foreach ($player as $row) {
+            $this->table->add_row($row);
         }
-        $this->data['trading_activity'] = $this->table->generate();
+
+        return $this->table->generate();
+    }
+
+    /**
+     * Generates the current holdings table with a given player
+     * @param $player the portfolio player
+     * @return mixed The generated table
+     */
+    function generateCurrentHoldingsTable($player) {
+        $this->table->set_heading('Stock Code','Cash', 'Date');
+        foreach ($player as $row) {
+            $this->table->add_row($row);
+        }
+
+        return $this->table->generate();
     }
 }